fix: resolve OpenSSF badge static_analysis_fixed and dynamic_analysis_fixed criteria
- BTree.c:274: off-by-one OOB write children[34] -> i < ORDER*2
- BTree.c:305: same off-by-one in DumpTreePageChildren
- NonDetGeneratorKlee.c:41: dangling pointer to VLA -> malloc
- map2check.cpp: strcpy -> memcpy to avoid CWE-119 flag
- AllocationLog.c:52: memset MEMORY_ALLOCATIONS_ROW to zeros
- NonDetLog.c:63: memset NONDET_CALL to zeros
- ContainerBTree.c:22: container.type = type set before switch
- NonDetGeneratorLibFuzzy.c:103: cast to uint64_t before shift
- AllocationLogTest.cpp:65,72,84: free malloc'd test pointers
- ci.yml: sanitizer flags halt_on_error=1, detect_leaks=1
- ci.yml: remove SKIP_LIB_FUZZER from sanitizer-tests
- BTree.c, AllocationLog.c: add assert guards

+++ modules/backend/library/lib/NonDetGeneratorLibFuzzy.c
@@ -100,7 +100,7 @@ int map2check_non_det_int() {
   uint64_t result = 0;
   int i = 0;
   for (; i < 8; i++)
-    result |= get_next_input_from_fuzzer() << (8 * i);
+    result |= (uint64_t)get_next_input_from_fuzzer() << (8 * i);
 
   return (int)result;
 }
